services/seismicdb/requirements.txt
minio
crate
msgpack-python
gunicorn
flask
flask-appconfig>=0.10
flask-api
requests